Feat: 검색화면 검색조건 별 가게 리스트 API MyBatis -> JPA 마이그레이션 작업

+++ src/main/java/com/support/matgo/store/dto/request/CoordinatesRequest.java
@@ -7,8 +7,13 @@ import java.math.BigDecimal;
 
 @Getter
 public class CoordinatesRequest {
+  /**
+   * 소수점 오차 범위
+   * 컴퓨터는 부동소수점 방식때문에 15자리부터 오차 발생
+   * 위/경도에서 5자리까지 약 1m, 6자리는 약 10cm, 7자리는 약 1cm의 차이
+   */
   @NotNull
-  private BigDecimal latitude;
+  private double latitude;
   @NotNull
-  private BigDecimal longitude;
+  private double longitude;
 }
